refactoring the codes
selecting the best code and adding all other codes into the trial codes repo
